postgresql-testing: Split into several files.
