+++ libs-scala/postgresql-testing/src/main/scala/com/digitalasset/testing/postgresql/PostgresAround.scala
@@ -9,81 +9,14 @@ import java.nio.charset.StandardCharsets
 import java.nio.file.{Files, Path, Paths}
 import java.util.concurrent.atomic.AtomicBoolean
 
-import com.digitalasset.daml.bazeltools.BazelRunfiles.rlocation
-import com.digitalasset.ledger.api.testing.utils.Resource
+import com.digitalasset.testing.postgresql.PostgresAround._
 import org.apache.commons.io.{FileUtils, IOUtils}
-import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach}
 import org.slf4j.LoggerFactory
 
 import scala.collection.JavaConverters.asScalaBufferConverter
 import scala.util.control.NonFatal
 
-trait PostgresAroundAll extends PostgresAround with BeforeAndAfterAll {
-  self: org.scalatest.Suite =>
-
-  override protected def beforeAll(): Unit = {
-    // we start pg before running the rest because _generally_ the database
-    // needs to be up before everything else. this is relevant for
-    // ScenarioLoadingITPostgres at least. we could much with the mixin
-    // order but this was easier...
-    startEphemeralPostgres()
-    super.beforeAll()
-  }
-
-  override protected def afterAll(): Unit = {
-    super.afterAll()
-    stopAndCleanUpPostgres()
-  }
-}
-
-trait PostgresAroundEach extends PostgresAround with BeforeAndAfterEach {
-  self: org.scalatest.Suite =>
-
-  override protected def beforeEach(): Unit = {
-    // we start pg before running the rest because _generally_ the database
-    // needs to be up before everything else. this is relevant for
-    // ScenarioLoadingITPostgres at least. we could much with the mixin
-    // order but this was easier...
-    startEphemeralPostgres()
-    super.beforeEach()
-  }
-
-  override protected def afterEach(): Unit = {
-    super.afterEach()
-    stopAndCleanUpPostgres()
-  }
-}
-
-case class PostgresFixture(
-    jdbcUrl: String,
-    port: Int,
-    tempDir: Path,
-    dataDir: Path,
-    confFile: Path,
-    logFile: Path,
-)
-
-private class PostgresResource extends Resource[PostgresFixture] with PostgresAround {
-
-  override def value: PostgresFixture = postgresFixture
-
-  override def setup(): Unit = {
-    startEphemeralPostgres()
-  }
-
-  override def close(): Unit = {
-    stopAndCleanUpPostgres()
-  }
-}
-
-object PostgresResource {
-  def apply(): Resource[PostgresFixture] = new PostgresResource
-}
-
 trait PostgresAround {
-
-  import PostgresAround._
-
   @volatile
   protected var postgresFixture: PostgresFixture = _
 
@@ -247,8 +180,6 @@ object PostgresAround {
 
   private val testUser = "test"
 
-  private val isWindows = sys.props("os.name").toLowerCase contains "windows"
-
   private def findFreePort(): Int = {
     val s = new ServerSocket(0)
     val port = s.getLocalPort
@@ -257,21 +188,6 @@ object PostgresAround {
     // turns out to be an issue, we need to find an atomic way of doing that.
     s.close()
     port
-  }
-
-  private case class Tool private[Tool] (name: String) {
-    import Tool._
-
-    def path: Path = rlocation(binPath.resolve(name + binExtension))
-  }
-
-  private object Tool {
-    private[Tool] val binPath = Paths.get("external", "postgresql_dev_env", "bin")
-    private[Tool] val binExtension = if (isWindows) ".exe" else ""
-
-    val createdb = Tool("createdb")
-    val initdb = Tool("initdb")
-    val pg_ctl = Tool("pg_ctl")
   }
 
   @SuppressWarnings(Array("org.wartremover.warts.Option2Iterable"))
